+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core.rs
@@ -1,0 +1,7 @@
+pub mod components;
+mod doenet_ml_core;
+pub mod dast;
+pub mod dependency;
+pub mod state;
+
+pub use self::doenet_ml_core::*;
