+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components.rs
@@ -1,0 +1,8 @@
+pub mod common;
+pub mod component_creation;
+pub mod doenet;
+pub mod special;
+pub mod preamble;
+
+pub use common::*;
+pub use special::*;
