+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/doenet.rs
@@ -1,0 +1,5 @@
+pub mod document;
+pub mod p;
+pub mod section;
+pub mod text;
+pub mod text_input;
