+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/doenet/section.rs
@@ -1,0 +1,12 @@
+use std::collections::HashMap;
+
+use crate::components::preamble::*;
+
+#[derive(Debug, Default, ComponentNode, RenderedComponentNode)]
+pub struct Section {
+    pub common: ComponentCommonData,
+}
+
+impl ComponentNodeStateVariables for Section {
+    fn initialize_state_variables(&mut self) {}
+}
