+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/special.rs
@@ -1,0 +1,2 @@
+pub mod _error;
+pub mod _external;
