+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components/special/_error.rs
@@ -1,0 +1,7 @@
+use crate::components::preamble::*;
+
+#[derive(Debug, Default, ComponentNode, ComponentNodeStateVariables, RenderedComponentNode)]
+pub struct _Error {
+    pub common: ComponentCommonData,
+    pub message: String,
+}
