+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/state/essential_state.rs
